pyproject.toml
[tool.poetry]
name = "fyp-langgraph"
version = "0.1.0"
description = ""
authors = ["Muhammad Abdullah <ma7800501@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.59"
langchain = "^0.3.12"
langchain-core = "^0.3.25"
langchain-openai = "^0.2.12"
langchain-groq = "^0.2.1"
python-pptx = "^1.0.2"
requests = "^2.32.3"
pydantic = "^2.10.3"
wikipedia = "^1.4.0"
langchain-community = "^0.3.12"
chromadb = "^0.5.23"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.3.1"
streamlit = "^1.41.1"
streamlit-carousel = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
